services/requirements.txt
Flask~=1.0.2
MarkupSafe~=1.1.0
objprint~=0.0.3
paho-mqtt~=1.5.1